repositories/voice-agent-3D/pyproject.toml
[tool.setuptools]
py-modules = ["agent", "tts_piper"]

[project]
name = "voice-agent-3d"
version = "0.1.0"
description = "3D Avatar voice agent - LiveKit + LemonSlice + Deepgram + ElevenLabs + Claude"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[anthropic,silero,mcp]~=1.5",
    "python-dotenv",
]

[project.optional-dependencies]
local = [
    "faster-whisper",
    "piper-tts",
    "numpy",
]
cloud = [
    "livekit-plugins-deepgram~=1.5",
    "livekit-plugins-elevenlabs~=1.5",
    "livekit-plugins-turn-detector~=1.5",
    "livekit-agents[lemonslice]",
]

[project.scripts]
start = "agent:server"
